Move rendering functions into separate modules
