+++ assets/src/scripts/index.spec.js
@@ -13,6 +13,9 @@ import 'ngwmn/components/graph/state/layout.spec.js';
 import 'ngwmn/components/graph/state/options.spec.js';
 import 'ngwmn/components/graph/state/points.spec.js';
 import 'ngwmn/components/graph/state/scales.spec.js';
+import 'ngwmn/components/graph/view/axes.spec.js';
+import 'ngwmn/components/graph/view/cursor.spec.js';
+import 'ngwmn/components/graph/view/water-levels.spec.js';
 import 'ngwmn/lib/ajax.spec';
 import 'ngwmn/lib/utils.spec';
 import 'ngwmn/services/cache.spec';
